Remove PuppetX::Firewalld::Property::RichRuleAction
This parent property was only used in one type for one parameter.
Moving the validation to the type fixes environment bleed caused by
loading pure ruby code like this.

Fixes #226
